### 2017-02-21 #175 @acabunoc
Hey @nuest & @desilinguist 👋 ! would either of you be interested in reviewing this paper? It looks like you both have CSS + R experience.

Also, @lebebr01 do you have any potential reviewers you can suggest? It's nice to have multiple reviewers

Review guidelines can be found here: http://joss.theoj.org/about#reviewer_guidelines
