### 2017-02-21 #175 @acabunoc
Thanks @lebebr01!

@daattali & @sckott - would love to have you as reviewers! For an idea of what you'd do, reviewer guidelines: http://joss.theoj.org/about#reviewer_guidelines
