### 2017-02-21 #175 @desilinguist
Sure I'd be happy to review :)
On Tue, Feb 21, 2017 at 5:04 PM Brandon LeBeau <notifications@github.com>
wrote:

> These two are not on the reviewer list, but perhaps @daattali
> <https://github.com/daattali> or @sckott <https://github.com/sckott>
> would be interested and I think would be good reviewers.
>
