### 2017-02-21 #185 @whedon
[REVIEW]: highlightHTML: CSS Formatting of R Markdown Documents
**Submitting author:** @lebebr01 (<a href="http://orcid.org/0000-0002-1265-8761">Brandon LeBeau</a>)
**Repository:** <a href="https://github.com/lebebr01/highlightHTML" target ="_blank">https://github.com/lebebr01/highlightHTML</a>
**Version:** v 0.1.3

- [x] **Version:** Does the release version given match the GitHub release (v 0.1.3)?
