### 2017-02-21 #175 @daattali
Hello! This package looks interesting, it's such a commonly needed tool.
Thanks for building it! I have a bit too much on my plate right now so as
terrible as this is for my PR, I'm going to stay out of the review process
and just wait to see it on CRAN :) Good luck!

---
Founder & Lead Consultant
AttaliTech Ltd
http://deanattali.com/shiny
