### 2017-03-03 #185 @desilinguist
It would be really useful if there was some example included in the repository and in the README that made it easy to try out this package. `vignette(highlightHTML)` also doesn't show me how to use this. What does the `tags` argument take in the `highlight_html()` function? The README just says that it takes a "Vector of CSS code" which is not very helpful.

It'd be really helpful if a usage example can be added so I can continue the review. Thanks!
