### 2017-03-06 f8b5cc8
Edit of README for JOSS.
