### 2017-03-06 #185 @lebebr01
Thanks for the comments. I agree, the README could have been better. This has been updated to reflect three things, installation is focused more on the CRAN version. Secondly, an example has been pulled from the vignette to remove the vague line about the tags argument. Third, description about installing the vignettes from the development version was added as well.

A second note, there is a vignette for this package, but these may not install directly from the development version on GitHub. You can view it directly on GitHub here: https://github.com/lebebr01/highlightHTML/tree/master/vignettes

You could also install from GitHub passing the build_vignettes argument: `devtools::install_github('lebebr01/highlightHTML', build_vignettes = TRUE)`.
