### 2017-08-09 #185 @lebebr01
@arfon and @desilinguist, I wanted to check in to see if there is any additional information I can provide to help with this? Are there still issues viewing the vignette or is the README example too simple to be helpful?
