### 2017-08-09 #185 @arfon
> Sorry, I have been swamped with other things. LGTM 👍

@desilinguist - thanks for getting back to us on this. There is still a substantial number of unchecked items above. Would you be able to update the checklist?
