### 2017-09-04 #185 @desilinguist
I really haven't had the time. Please accept my sincere apologies. Given
what the rest of my year looks like, it might indeed be prudent to look for
another reviewer.
On Sun, Sep 3, 2017 at 20:09 Abigail Cabunoc Mayes <notifications@github.com>
wrote:

> Pinging @desilinguist <https://github.com/desilinguist>, have you had
> time to go through the checklist?
>
> Let me know if you're too busy and I can look for another reviewer 😄
>
