### 2017-09-09 #185 @acabunoc
@daattali here are the reviewer guideline: http://joss.theoj.org/about#reviewer_guidelines

But mostly, it's going through the checklist at the top of this issue.

Thanks again! 🎉 🎈 🍰
