### 2017-10-18 #185 @daattali
Review submitted https://github.com/lebebr01/highlightHTML/issues/5
