### 2017-10-23 #185 @lebebr01
Thank you, appreciate the review and thoughtful comments. I'll work through these and post a response/changes in the coming weeks.
