### 2017-11-29 fc1cf86
Improved documentation based on JOSS review
