+++ R/util.r
@@ -1,12 +1,12 @@
-update_css <- function(x, tags) {
+update_css <- function(html_file, tags) {
   
-  location <- max(grep("</style>", x, ignore.case = TRUE))
+  location_css <- max(grep("</style>", html_file, ignore.case = TRUE))
   
   # # split file
-  tmpA <- x[1:(location - 1)]
-  tmpB <- x[location:length(x)]
+  html_preamble <- html_file[1:(location_css - 1)]
+  html_main <- html_file[location_css:length(html_file)]
   
   ## adding custom CSS
-  tmp <- c(tmpA, tags, tmpB)
-  tmp
+  updated_html <- c(html_preamble, tags, html_main)
+  updated_html
 }
