+++ tests/testthat/test_structure.r
@@ -1,4 +1,4 @@
-context('test structure')
+context('test tag insertion')
 
 test_that('correct print structure', {
   file <- system.file('examples', 'bgtext.html', package = 'highlightHTML')
@@ -9,7 +9,14 @@ test_that('correct print structure', {
             "#colgreen {color: green;}")
   
   # Post-process HTML file
-  tmp <- highlight_html(input = file, output = tempfile(fileext = ".html"), update_css = TRUE, 
-                 tags = tags, browse = FALSE, print = TRUE)
-  expect_true(is.character(tmp))
+  text_output <- highlight_html(input = file, output = tempfile(fileext = ".html"), 
+                 tags = tags, browse = FALSE, print = TRUE, render = FALSE)
+  
+  expect_true(any(grepl("#bgblack", text_output)))
+  expect_true(any(grepl("#bgblue", text_output)))
+  expect_true(any(grepl("#colgreen", text_output)))
+  
+  expect_true(any(grepl("id='bgblack'", text_output)))
+  expect_true(any(grepl("id='bgblue'", text_output)))
+  expect_true(any(grepl("id='colgreen'", text_output)))
 })
