### 2017-11-29 eee2019
Updated review response and switched warning to message
