+++ inst/joss/reviewer_response.md
@@ -5,13 +5,12 @@ Thanks @daattali for the thoughtful comments and suggestions. In general the fol
 
 ## README
 
-- First 6 bullets are related and bundled here. In general the README has been reworked significantly. The introduction to the package has been clarified and expanded slightly to better discuss what I view is a typical use of the package. Some of this has been borrowed from the paper.md JOSS submission. The trivial example has been replaced with a more typical workflow from Rmd to HTML directly with the functions in the package, this is shown in the section "Simple Example of Package Workflow". In the example, I show the raw input Rmd file, the look of the intermediate markdown table with included CSS ids, and then the final code step to compile the document into an HTML. Finally, a screenshot is included to show the look of the final output. I also attempted to change "table" in a few places to "summary table" to reflect the type of table that I feel should be passed to the `table_id_inject` function.
+- First 6 bullets are related and bundled here. In general the README has been reworked significantly. The introduction to the package has been clarified and expanded slightly to better discuss what I view is a typical use of the package. Some of this has been borrowed from the paper.md JOSS submission. The trivial example has been replaced with a more typical workflow from Rmd to HTML directly with the functions in the package, this is shown in the section "Simple Example of Package Workflow". In the example, I show the raw input Rmd file, the look of the intermediate markdown table with included CSS ids, and then the final code step to compile the document into an HTML. Finally, a screenshot is included to show the look of the final output. I also attempted to change "table" in a few places to "summary table" to reflect the type of table that I feel should be passed to the `table_id_inject` function. I'd be interested in feedback on whether this is both easier to read and offers a more compelling use/workflow for the package.
 - Added line to README about installation of devtools package.
 - Fixed `library(hightlightHTML)` typo.
 - The example with `bgtable.html` was removed as this was a contrived example, primarily for testing. A newer example from Rmd to HTML was created to reflect the most useful use case for the package. This file is called `joss.Rmd`.
-- The code should be `update_css`. This typo has been fixed from an older version of the package.
-- `update_css = FALSE` is mainly used internally and as such has been removed from the master `highlight_html` function, instead just passed internally to the two `highlight_html_cells` and `highlight_html_text` functions.
-- Great suggestion on not overwriting input file. This has been remedied to include a warning if the output file location is not specified and the output is then "<input>_out.html".
+- `update_css = FALSE` is used internally and as such has been removed from the master `highlight_html` function, instead just passed internally to the two `highlight_html_cells` and `highlight_html_text` functions. Discussion of this argument has also been removed from the README and vignette.
+- Great suggestion on not overwriting input file. This has been remedied to include a message if the output file location is not specified and the output is then "<input>_out.html". The file path of the generated output file is included in the printed message.
 - Updated package DESCRIPTION to better reflect package goals and maintain consistency with README documentation.
 
 ## Vignette
@@ -22,11 +21,11 @@ Thanks @daattali for the thoughtful comments and suggestions. In general the fol
 
 ## Code
 
-- Included better/more tests. This is an area I have improved on since this package was frozen for JOSS review.
+- Adjusted tests. This is an area I have commonly struggled with implementation. I believe the tests are a bit more helpful to help distinguish whether specific features are working. I am now testing whether the CSS tags and ids are found within the output document. I'm open to any additional tests you think about be useful here.
 - Internal names were replaced to more descriptive names to help improve debugging in the future. Very useful suggestion.
 - Adjusted `rgb2hex` function documentation to reflect named vs unnamed list specification.
 - Added error handling to catch double NULL case for `rgb2hex` function.
 - Switched code where relevant to `&&` instead of `&`. I wasn't aware of the difference actually. Thanks for pointing this out.
[…]
