### 2017-11-29 e6e6783
Updated news and addes news.md to buildignore
