### 2017-11-30 c611f1c
Update to README/vignette to improve readability
