+++ vignettes/the_basics.utf8.md
@@ -1,7 +1,7 @@
 ---
 title: "The Basics of highlightHTML"
 author: "Brandon LeBeau"
-date: "`r Sys.Date()`"
+date: "2017-11-30"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Vignette Title}
@@ -36,7 +36,8 @@ The addition of the *#bgblue* and *#bgred* indicates which cells will be changed
 
 The function to use for the post-processing is `highlight_html` and requires three arguments, the input file, the output file, and the CSS tags themselves. This will look something like the following using an example file from the package:
 
-```{r function, eval=FALSE}
+
+```r
 library(highlightHTML)
 file <- system.file('examples', 'bgtable.html', 
                     package = 'highlightHTML')
@@ -61,7 +62,8 @@ Styling text is also simple in a markdown document. Braces `{}` are used to iden
 For example, perhaps there is `{#colgold some text}` to turn gold. In the previous example, the text "some text" would be processed to be styled using the CSS id #colgold. 
 
 Another example file can be run with the following commands to show more text styling examples:
-```{r text, eval = FALSE}
+
+```r
 file <- system.file('examples', 'bgtext.html', package = 'highlightHTML')
 
 # Change background color and text color with CSS
@@ -96,7 +98,7 @@ output: html_document
 ## Simple Markup
 The `highlightHTML` packages allows for {#bggrey simple markup} to add styling to text and tables. Using the `rmarkdown` package and the `table_id_inject` function allows users to easily add markup to Rmd documents and render directly to HTML.
 
-```{r symm, echo = FALSE, results = 'asis', message = FALSE}`r ''`
+```{r symm, echo = FALSE, results = 'asis', message = FALSE}
 library(dplyr)
 library(highlightHTML)
 
@@ -114,27 +116,22 @@ chickwts %>%
 ````
 
 The resulting summary table would look like the following in markdown:
-```{r symm, echo = FALSE, results = 'asis', message = FALSE}
-library(dplyr)
-library(highlightHTML)
 
-chickwts %>%
-  group_by(feed) %>%
-  summarise(avg_weight = mean(weight),
-            sd_weight = sd(weight)) %>%
-  mutate(feed = as.character(feed)) %>%
-  table_id_inject(id = c('#bggrey', '#bgblack', '#bglightred', '#textblue'), 
-                conditions = c('> 270', '> 300', '> 60', '== "horsebean"'),
-                variable = list('avg_weight', 'avg_weight', 'sd_weight', 'feed'),
-                num_digits = 3) %>%
-  knitr::kable(format = 'markdown')
-```
+|feed                |avg_weight       |sd_weight          |
+|:-------------------|:----------------|:------------------|
+|casein              |323.583 #bgblack |64.434 #bglightred |
+|horsebean #textblue |160.2            |38.626             |
+|linseed             |218.75           |52.236             |
+|meatmeal            |276.909 #bggrey  |64.901 #bglightred |
+|soybean             |246.429          |54.129             |
+|sunflower           |328.917 #bgblack |48.836             |
 
[…]
