### 2017-11-30 #185 @lebebr01
Thanks again @daattali for the review. I posted my comments on edits made here: https://github.com/lebebr01/highlightHTML/issues/5. As mentioned in that issue, most of the changes I grouped into a single commit: https://github.com/lebebr01/highlightHTML/commit/fc1cf86e2aff8a14f986bf9c7520e313611dbc78.

I'd be very interested on your thoughts on the changes. Thanks again in advance.
