+++ contributing.md
@@ -1,0 +1,21 @@
+# Contributing to highlightHTML
+
+## Issues/Bugs
+
+Please report bugs and contribute feature requests using the the Github Issues page. For reporting bugs, please provide a reproducible example when possible. If not possible, please provide as much information as possible to aid in recreation of the bug. For feature requests, please give an example use case.
+
+## Development guidelines
+
+New features or improvements to the highlightHTML code or documentation are welcomed. Please use GitHub's pull request feature for proposing and submitting changes.  
+
+New features or changes to existing features should mimic the style used for the highlightHTML package. The package style currently uses "_" between function words and aims for names to user functions to be descriptive without being excessively long.
+
+If adding a new feature, a test should be added for that new feature, as well as an update to the vignette to document the new feature. Testing is done using the testthat package. If you have not written a test before, please submit example code with a clear explanation of the intended output from the example code.
+
+## Questions about these guidelines?
+
+Please use the Issues page for questions about these guidelines and please label the issue with the "questions" label. 
+
+## Code of Conduct
+
+When contributing to highlightHTML, you are expected to follow the [code of conduct](https://github.com/lebebr01/highlightHTML/blob/master/conduct.md).
