### 2018-01-10 lebebr01/highlightHTML#5 @daattali
This version is much better!

A few comments:

- In the README, the command to install the package from devtools is `devtools::install_github('lebebr01/highlightHTMl')` instead of `devtools::install_github('lebebr01/highlightHTML')` (the last "l" in the package name is not capitalized. This will technically still work but probably better to fix it)

- I would put the example of `{#bggrey example}` inside an inline code block, so that it's clear what text exactly is example code vs explanation text.

- I just realized something: CSS IDs are meant to only be applied to one element. IDs are supposed to be unique. But in the examples you show, there are multiple cells that you're giving the same ID to. To me it sounds like a CSS class rather than an ID is more fitting.

- Simple example in the README: It's definitely much better, but it still takes a lot of brain power from the reader to understand what's going on. I think it would make it significantly clearer if you show what the original table looks like, then you say what hashtags you'll add to what cells (for example, say that you'll add "#bgblack" to any cell in the weight column that is larger than 300, #bggrey to any weight over 270, etc) (like the basic example in rmd). The "basic example" in the vignette is great - that's exactly what I'd want to see initially, it explains exactly what the input is, what you need to do , and what the result is

- The arguments to `table_id_inject()` have to be the same length, right? Then there should be some check for that, with an appropriate error message. I tried removing one element from the `variable` list and I got a generic error.

- Testing: much better! I would remove the one test that is commented out, or fix it. And I wouldn't use dplyr inside tests, I would make them as dependency-free as possible, to make sure they only test your code. Unit tests should be as small and independent as possible.

These are all fairly small changes compared to what I reviewed last time, so no need to let me know when you do it, I trust that you'll implement them. The package looks good now :+1:
