### 2018-01-10 #185 @daattali
After revisions, the package LGTM

I left a few more small comments for the author to fix, but they are not blocking issues, I would accept the package now
