### 2018-01-10 lebebr01/highlightHTML#5 @lebebr01
Thanks @daattali. I appreciate your comments and you taking the time to look through the package. I'll implement these soon.

Cheers.
