### 2018-01-11 9f87d18
Updates based on reviewer comments #5
