+++ tests/testthat/test_table_id.r
@@ -27,8 +27,14 @@ test_that('add in tags to table', {
                          variable = list(c('sd_mpg'), c('avg_mpg', 'sd_mpg')))
   
   expect_true(grepl('#bgblue', inject_table['avg_mpg']))
-  #expect_false(grepl('#bgred', tmp['sd_mpg']))
   expect_true(grepl('#bgblue', inject_table['sd_mpg']))
   
+  expect_error(table_id_inject(summ_table, id = c('#bgred', '#bgblue'), 
+                               conditions = c('< 2'),
+                               variable = list(c('sd_mpg'), c('avg_mpg', 'sd_mpg'))))
+  
+  expect_error(table_id_inject(summ_table, id = c('#bgred', '#bgblue'), 
+                               conditions = c('< 2', '< 16'),
+                               variable = list(c('avg_mpg', 'sd_mpg'))))
   
 })
