+++ vignettes/the_basics.utf8.md
@@ -1,152 +1,0 @@
----
-title: "The Basics of highlightHTML"
-author: "Brandon LeBeau"
-date: "2017-11-30"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Vignette Title}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-Have you ever wanted to include some form of styling, either conditional or nonconditional, when going from a (R)markdown input document to a html output document? If so, this package may be worthwhile to you. The `highlightHTML` package inject CSS into the output HTML document automatically via CSS ids (more information on CSS ids here: <http://www.cssbasics.com/css-ids/>). 
-
-The examples below start by showing the syntax in somewhat trivial examples initially. This is done to help show the markdown syntax a bit more simply. Toward the end of the document, a more typical workflow going from Rmd to HTML is shown that puts all the pieces together.
-
-# Basic Example
-Suppose you have a summary table like the following (note: this table was created by typing manually, a more typical workflow is shown later using R directly):
-
-| Color Name | Number       |
-|----------- | ------------ |
-| Red        |  25          |
-| Blue       |  40          |
-| Green      |  4           |
-| Orange     |  87          |
-
-You could then add styling by adding the following tags to the table, again this was added by hand, but a conditional example with R is shown later.
-
-| Color Name | Number       |
-|----------- | ------------ |
-| Red        |  25  #bgred  |
-| Blue       |  40  #bgblue |
-| Green      |  4           |
-| Orange     |  87          |
-
-The addition of the *#bgblue* and *#bgred* indicates which cells will be changed. After turning the markdown document into an html file, this package can now be used to post-process the html file. The post-processing will add an id value for each cell with the *#bgblue* or *#bgred* and remove those from the table. 
-
-The function to use for the post-processing is `highlight_html` and requires three arguments, the input file, the output file, and the CSS tags themselves. This will look something like the following using an example file from the package:
-
-
-```r
-library(highlightHTML)
-file <- system.file('examples', 'bgtable.html', 
-                    package = 'highlightHTML')
-tags <- c("#bgred {background-color: #FF0000;}", 
-  "#bgblue {background-color: #0000FF;}")
-highlight_html(input = file, 
-               output = tempfile(fileext = ".html"), 
-               tags = tags,
-               browse = TRUE,
-               print = FALSE, 
-               render = FALSE)
-```
-
-In the above `highlight_html` function call, the output file is saved to a temporary file. The `update_css` argument will inject the tags into the html document within any style tags that are found within the document. The `browse` argument will open the resulting post-processed file in the default browser to be viewed directly. Finally, the `print` argument if TRUE, will print the html output to the R console. This may be useful for verification of potential errors in the post-processing. The `render` argument is set to FALSE here (TRUE is the default) as the input file was already an HTML file. 
-
-![bgtable](bgtable_output.png)
-
-
-# Styling Text
-Styling text is also simple in a markdown document. Braces `{}` are used to identify text that should be styled A CSS id tag is then added to link this to the CSS styling 
-
-For example, perhaps there is `{#colgold some text}` to turn gold. In the previous example, the text "some text" would be processed to be styled using the CSS id #colgold. 
-
-Another example file can be run with the following commands to show more text styling examples:
-
-```r
-file <- system.file('examples', 'bgtext.html', package = 'highlightHTML')
-
-# Change background color and text color with CSS
-tags <- c("#bgblack {background-color: black; color: white;}", 
-  "#bgblue {background-color: #0000FF; color: white;}",
-  "#colgreen {color: green;}")
-  
-# Post-process HTML file
[…]
