### 2018-01-11 3e8319f
Update news.md for 0.2.0
