### 2018-01-11 14c8408
Bump version with zenodo metadata
